Create Corners.java
Corners is a game played in Eastern states of India.
It’s usually played with 4 or more players, where each player occupies a spot, representing corners of a polygon. A player who’s chosen to be the Tat must go to a player at one of the corners, who will then redirect him/her to one of the other corners. While the Tat is on his/her way to the redirected corner, others try to exchange their spot and Tat would try to take it as an opportunity to capture the corner, leaving one of the others as the new Tat, similar to Musical Chairs.
Mili was one day playing Corners with her friends, and a thought struck her mind. If all the players holding the corners make up their mind, and have decided where to redirect the Tat, then it may be possible that the Tat is stuck in a loop (or cycle) indefinitely.
Find the loop having the maximum number of corners.
Note: A player may redirect the Tat to himself/herself, representing a loop involving a single corner.

Input Format
First line contains an integer N, denoting the number of corners.
Next line contains N space separated integers, from the range [0, N-1], where each integer represents the index (or position), where that player would redirect the Tat.
Note: Each value is unique.

Constraints
1 <= N <= 105
1 <= Ai < N

Output Format
Print the number of corners in the longest loop.

Sample Input
7
5 4 0 3 1 6 2
Sample Output
4
Explanation
The player at index 0 would redirect the Tat to index 5 (A0 = 5), where he/she would be redirected to index 6 (A5 = 6). Then from index 6, the Tat would be redirected to index 2 (as A6 = 2). From the index 2, he/she would be redirected to index 0 (A2 = 0) bringing him to the place he started.
The loop forms a cycle as, 0 -> 5 -> 6 -> 2 and then back to 0. The length is 4.